Moved platform-specific helper functions to separate module.
