+++ integration_tests/platforms/platform_sh/test_deployment.py
@@ -3,52 +3,7 @@ import re, time
 import pytest
 
 from ...utils import it_helper_functions as it_utils
-from ...utils.it_helper_functions import make_sp_call
-
-
-# --- Platform-specific helper functions ---
-
-def create_platformsh_project():
-    """Create a project on Platform.sh."""
-    print("\n\nCreating a project on Platform.sh...")
-    org_output = make_sp_call("platform org:info", capture_output=True).stdout.decode()
-    org_id = re.search(r'([A-Z0-9]{26})', org_output).group(1)
-    print(f"  Found Platform.sh organization id: {org_id}")
-    make_sp_call(f"platform create --title my_blog_project --org {org_id} --region us-3.platform.sh --yes")
-
-def commit_configuration_changes():
-    """Commit configuration changes made by simple_deploy."""
-    print("\n\nCommitting changes...")
-    make_sp_call("git add .")
-    make_sp_call("git commit -am 'Configured for deployment.'")
-
-def push_project():
-    """Push a non-automated deployment."""
-    # Try pausing before making push.
-    time.sleep(30)
-
-    print("Pushing to Platform.sh...")
-    make_sp_call("platform push --yes")
-
-    project_url = make_sp_call("platform url --yes", capture_output=True).stdout.decode().strip()
-    print(f" Project URL: {project_url}")
-
-    project_info = make_sp_call("platform project:info", capture_output=True).stdout.decode()
-    project_id = re.search(r'\| id             \| ([a-z0-9]{13})', project_info).group(1)
-    print(f"  Found project id: {project_id}")
-
-    return project_url, project_id
-
-def get_project_url_id():
-    """Get project URL and id of a deployed project."""
-    project_info = make_sp_call("platform project:info", capture_output=True).stdout.decode()
-    project_id = re.search(r'\| id             \| ([a-z0-9]{13})', project_info).group(1)
-    print(f"  Found project id: {project_id}")
-
-    project_url = make_sp_call("platform url --yes", capture_output=True).stdout.decode().strip()
-    print(f" Project URL: {project_url}")
-
-    return project_url, project_id
+from . import platformsh_helper_functions as platformsh_utils
 
 
 # --- Test functions ---
@@ -69,23 +24,23 @@ def test_platformsh_deployment(tmp_project, cli_options):
 
     # Create a new project on the remote host, if not testing --automate-all.
     if not cli_options.automate_all:
-        create_platformsh_project()
+        platformsh_utils.create_platformsh_project()
 
     # Run simple_deploy against the test project.
     it_utils.run_simple_deploy(python_cmd, 'platform_sh', cli_options.automate_all)
 
     # If testing Pipenv, lock after adding new packages.
     if cli_options.pkg_manager == 'pipenv':
-        make_sp_call(f"{python_cmd} -m pipenv lock")
+        it_utils.make_sp_call(f"{python_cmd} -m pipenv lock")
 
     # Get the deployed project's URL, and ID so we can destroy it later.
     #   This also commits configuration changes and pushes the project
     #   when testing the configuration-only workflow.
     if cli_options.automate_all:
-        project_url, project_id = get_project_url_id()
+        project_url, project_id = platformsh_utils.get_project_url_id()
     else:
-        commit_configuration_changes()
-        project_url, project_id = push_project()
+        platformsh_utils.commit_configuration_changes()
+        project_url, project_id = platformsh_utils.push_project()
 
     # Test functionality of both deployed app, and local project.
     #   We want to make sure the deployment works, but also make sure we haven't
@@ -116,4 +71,4 @@ def test_platformsh_deployment(tmp_project, cli_options):
     if tear_down:
         print("\nCleaning up:")
         print("  Destroying Platform.sh project...")
-        make_sp_call(f"platform project:delete --project {project_id} --yes")
+        it_utils.make_sp_call(f"platform project:delete --project {project_id} --yes")
